requirements.txt
Flask==3.0.3
rembg==2.0.57
pillow==10.4.0